test and edit the CRUD category

+++ src/Form/Category/DeleteCategoryForm.php
@@ -36,6 +36,13 @@ class DeleteCategoryForm extends ABaseForm
     {
         $categoryId = self::getParams($request)['route']['id'];
 
+        if($this->categoryRepository->find($categoryId)
+                ->getShop()
+                ->getUser()
+                ->getId() !== $this->getUser()->getId()) {
+            throw new NotFoundHttpException('you are not allowed to delete this category');
+        }
+
         try {
             $this->categoryRepository->removeById($categoryId);
         } catch (EntityNotFoundException) {
